chatbot/requirements.txt
# Multi-Agent Chatbot System Dependencies

# Core AMP dependencies (local shared library)
-e ../shared-lib

# LangChain and AI dependencies
langchain>=0.0.350
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.3.0
tiktoken>=0.5.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Web interface and API
fastapi>=0.101.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiohttp>=3.8.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0

# Utilities and configuration
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.4.0
typing-extensions>=4.7.0

# Async and concurrent processing
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and observability (optional)
prometheus-client>=0.17.0
structlog>=23.1.0

# Production deployment
gunicorn>=21.2.0
redis>=4.6.0

# Optional integrations (uncomment as needed)
# salesforce-bulk>=2.2.0    # CRM integration
# jira>=3.5.0              # Ticket system integration
# slack-sdk>=3.21.0        # Slack notifications
# discord-py>=2.3.0        # Discord integration
# python-telegram-bot>=20.5  # Telegram integration
# sentry-sdk>=1.30.0       # Error tracking
# celery>=5.3.0            # Background task processing